move merkle tree into new module
